requirements.txt
Flask==2.3.3
pyOpenSSL==23.2.0
pyinstaller==5.13.2
